Converts test suites to use inspec, fixes errors
The Minitest Chef handler has been
[deprecated](https://github.com/chef/minitest-chef-handler#deprecation-notice)
and users are encouraged to switch to Inspec. This change removes the
usage of minitest from this project and updates the tests and verifier
to use Inspec instead. In addition, I discovered a few issues with
getting the tests to work:

* All test kitchen runs failed early because there was no attribute set
  for accepting the license to use wlp. This attribute is now set for
  all suites in the `.kitchen.yml`.
* The custom test suite failed to install the extras package because it
  doesn't exist on remote repository. I had to rename this to be
  `developers-extras`.
* The `install_feature` resource in the custom test recipe was using a
  property called `location`, but there is no property defined in
  `resources/install_feature.rb`, so I removed this usage.
* The installUtility wasn't able to find a feature called `mongodb` so I
  renamed it to `mongodb-2.0`.

All of the test cases are passing across all suites and platforms. If
for any reason the above issues are an oversight in my understanding of
how to make the tests work, please let me know and I can back pedal
these changes out.

Signed-off-by: Keith Walters <keith.walters@cattywamp.us>