Fixes validations in Chef 13 for LWRP attributes
The `:kind_of` types set for several LWRPs are Strings but their default
values are `nil`. In Chef13, there is apparently a strict validation
enforcing the types of the assigned values or default values of an
attribute match their type. The affected resources and attributes are:

* `install_feature` resource - `name` property
* `download_feature` resource - `name` and `directory` properties

The default values for `name` properties have been removed since names
are required for resources anyway, they should never be `nil`. The
`directory` property is now defaulted to the `node[:wlp][:base_dir]`
since that is what has happening behind the scenes in the provider
anyway.

Fixes #51

Signed-off-by: Keith Walters <keith.walters@cattywamp.us>