Parameterizes ownership of repo config in LWRPs
In the `install_feature` and `download_feature` resources, the step
setup the `repository.propertie` file was hard coding the owner and
group, whereas other config files used the `node[:wlp][:user]` and
`node[:wlp][:group]` attributes. This causes a problem if a wrapping
cookbook tries to change the user or group names, because they will
still have to have the hard coded values as a user or group on their
system, even if they don't want it. This just keeps everything
consistent.

Signed-off-by: Keith Walters <keith.walters@cattywamp.us>